demo: Keep getSiteConfigs and getSiteConfigForDomain in siteConfig.ts
The move into a separate module was meant to keep next/headers out of the
browser bundle, but that was never a risk: the NewsListBlock loader chain runs
exclusively in server components, and the block imports the loader type-only.
Reverting the move keeps this pull request to the structured-data change.
